Bugfix ferite_loworderindex breaks when shiftAmount is 2
Buggy behavior is shown by ferite_amt-hash-index-from-hash.c

The buggy code was:

    return (((index << (shiftAmount - AMT_SHIFT_AMOUNT)) >> 27) & 0x1F);

where shiftAmount is unsigned int and AMT_SHIFT_AMOUNT is 5.

The function was called successively with shiftAmount set to 32, 27, 22,
17, 12, 7, 2 (minus AMT_SHIFT_AMOUNT each time).

When shiftAmount is less than AMT_SHIFT_AMOUNT,

    shiftAmount - AMT_SHIFT_AMOUNT

becomes:

    (unsigned int)(2) - 5

which is a small unsigned integer that underflow to a very large
unsigned integer hence the buggy behavior.

Note that this bug has no effect on the behavior of the AMT as the value
is expected to be any value within 0-31 - the only visible effect is in
the actual placement of the hash items in memory.